hud_speedometer: int instead of uint16_t

+++ dmc/cl_dll/hud_speedometer.h
@@ -3,7 +3,7 @@
 
 class CHudSpeedometer : public CHudBase
 {
-	uint16_t speed;
+	int speed;
 
 	cvar_t* hud_speedometer;
